feat(example-validation-app): refactor the code to use middleware
